app/requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
attrs==19.1.0
Automat==0.7.0
certifi==2018.8.24
cffi==1.12.3
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
cssselect==1.0.3
defusedxml==0.5.0
Django==2.0
django-allauth==0.37.1
django-cors-headers==2.4.0
django-js-asset==1.1.0
django-mptt==0.9.1
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
docutils==0.14
dodgy==0.1.9
environ==1.0
flake8==3.5.0
flake8-polyfill==1.0.2
hyperlink==19.0.0
idna==2.6
incremental==17.5.0
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.3.3
Markdown==3.1
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.1.0
parsel==1.5.1
pep8-naming==0.7.0
prospector==1.1.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.3.1
pycparser==2.19
PyDispatcher==2.0.5
pydocstyle==2.1.1
pyflakes==1.6.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pyOpenSSL==19.0.0
pypiwin32==223
python-scrapyd-api==2.1.2
python3-openid==3.1.0
pytz==2018.5
pywin32==224
PyYAML==3.13
queuelib==1.5.0
requests==2.18.4
requests-oauthlib==1.0.0
requirements-detector==0.6
Scrapy==1.5.1
scrapy-djangoitem==1.1.1
scrapyd==1.2.0
service-identity==18.1.0
setoptconf==0.2.0
six==1.11.0
snowballstemmer==1.2.1
Twisted==18.9.0
typed-ast==1.1.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
virtualenv==16.0.0
w3lib==1.20.0
wrapt==1.10.11
zope.interface==4.6.0
